feat(nvim)!: use mini deps instead of lazy
