+++ config/nvim/after/plugin/comment.lua
@@ -1,0 +1,16 @@
+require("ts_context_commentstring").setup({
+  enable_autocmd = false,
+  custom_calculation = function(_node, language_tree)
+    if vim.bo.filetype == "blade" and language_tree._lang ~= "javascript" then
+      return "{{-- %s --}}"
+    end
+  end,
+})
+
+local get_option = vim.filetype.get_option
+
+---@diagnostic disable: duplicate-set-field
+vim.filetype.get_option = function(filetype, option)
+  return option == "commentstring" and require("ts_context_commentstring.internal").calculate_commentstring()
+    or get_option(filetype, option)
+end
