+++ config/nvim/init.lua
@@ -1,1 +1,5 @@
-require("config")
+-- require("config")
+vim.g.mapleader = vim.keycode("<space>")
+vim.g.maplocalleader = vim.keycode("<space>")
+
+require("manager")
