--- config/nvim/lua/plugins/autopairs.lua
@@ -1,26 +1,0 @@
-return {
-  "windwp/nvim-autopairs",
-  event = "InsertEnter",
-  config = function()
-    require("nvim-autopairs").setup({
-      check_ts = true,
-      ts_config = {
-        lua = { "string", "source" },
-        javascript = { "string", "template_string" },
-        java = false,
-      },
-      disable_in_visualblock = true,
-      fast_wrap = {
-        map = "<M-e>",
-        chars = { "{", "[", "(", '"', "'" },
-        pattern = string.gsub([[ [%'%"%)%>%]%)%}%,] ]], "%s+", ""),
-        offset = 0,
-        end_key = "$",
-        keys = "qwertyuiopzxcvbnmasdfghjkl",
-        check_comma = true,
-        highlight = "PmenuSel",
-        highlight_grey = "LineNr",
-      },
-    })
-  end,
-}
