--- config/nvim/lua/plugins/colors.lua
@@ -1,5 +1,0 @@
-return {
-  { "maxmx03/solarized.nvim" },
-  { "nyoom-engineering/oxocarbon.nvim" },
-  { "https://github.com/Skardyy/makurai-nvim" },
-}
