--- config/nvim/lua/plugins/conform.lua
@@ -1,37 +1,0 @@
-return {
-  "stevearc/conform.nvim",
-  init = function()
-    vim.o.formatexpr = "v:lua.require'conform'.formatexpr()"
-  end,
-  config = function()
-    require("conform").setup({
-      formatters_by_ft = {
-        nix = { "nixfmt" },
-        javascript = { "prettier" },
-        javascriptreact = { "prettier" },
-        typescript = { "prettier" },
-        typescriptreact = { "prettier" },
-        json = { "prettier" },
-        jsonc = { "prettier" },
-        scss = { "prettier" },
-        css = { "prettier" },
-        lua = { "stylua" },
-        php = { "php_cs_fixer" },
-        blade = { "blade-formatter" },
-        yaml = { lsp_format = "fallback" },
-        helm = { lsp_format = "fallback" },
-        ["_"] = { "trim_whitespace", "trim_newlines" },
-      },
-      format_after_save = {
-        lsp_fallback = false,
-        async = true,
-        timeout_ms = 1500,
-      },
-      formatters = {
-        prettier = {
-          require_cwd = true,
-        },
-      },
-    })
-  end,
-}
