--- config/nvim/lua/plugins/ui.lua
@@ -1,7 +1,0 @@
-return {
-  { "nvim-tree/nvim-web-devicons" },
-  {
-    "brianhuster/live-preview.nvim",
-    opts = { dynamic_root = true, picker = nil },
-  },
-}
